Flatten UI::Chart::AsyncFrame to UI::ChartAsyncFrame
