+++ app/views/organized/graduated_notifications/_results.html.erb
@@ -1,5 +1,5 @@
 <% if @render_chart && @available_graduated_notifications.present? %>
-  <%= render(UI::Chart::AsyncFrame::Component.new(
+  <%= render(UI::ChartAsyncFrame::Component.new(
     id: :graduated_notifications_chart_frame,
     src: organization_graduated_notifications_path(sortable_search_params.merge(organization_id: current_organization.to_param, chart_only: "1"))
   )) %>
