+++ app/views/organized/graduated_notifications/chart.html.erb
@@ -1,4 +1,4 @@
-<%= render(UI::Chart::AsyncFrame::Component.new(id: :graduated_notifications_chart_frame)) do %>
+<%= render(UI::ChartAsyncFrame::Component.new(id: :graduated_notifications_chart_frame)) do %>
   <% if @available_graduated_notifications.present? %>
     <%= render(UI::Chart::Component.new(
       series: time_range_counts(collection: @available_graduated_notifications.unscope(:order), column: "graduated_notifications.created_at"),
