+++ app/views/organized/impound_records/_results.html.erb
@@ -14,7 +14,7 @@
 <% end %>
 
 <% if @render_chart %>
-  <%= render(UI::Chart::AsyncFrame::Component.new(
+  <%= render(UI::ChartAsyncFrame::Component.new(
     id: :impound_records_chart_frame,
     src: organization_impound_records_path(sortable_search_params.merge(organization_id: current_organization.to_param, chart_only: "1"))
   )) %>
