+++ app/views/organized/impound_records/chart.html.erb
@@ -1,4 +1,4 @@
-<%= render(UI::Chart::AsyncFrame::Component.new(id: :impound_records_chart_frame)) do %>
+<%= render(UI::ChartAsyncFrame::Component.new(id: :impound_records_chart_frame)) do %>
   <% if available_impound_records.present? %>
     <%= render(UI::Chart::Component.new(series: time_range_counts(collection: available_impound_records), time_range: @time_range, stacked: true)) %>
   <% end %>
