+++ app/views/organized/registrations/_results.html.erb
@@ -1,5 +1,5 @@
 <% if @render_chart && @available_bikes.present? %>
-  <%= render(UI::Chart::AsyncFrame::Component.new(
+  <%= render(UI::ChartAsyncFrame::Component.new(
     id: :registrations_chart_frame,
     src: organization_registrations_path(sortable_search_params.merge(organization_id: current_organization.to_param, chart_only: "1"))
   )) %>
